# make getindex shorter and faster